Migrated to using through-bore encoder for elevator
Ready to test. PID may need to be retuned (may be too high, as it was tuned for internal encoder at the start of the gearbox reduction). Chose to only zero external encoder when limit switch is pressed, since it theoretically should be more accurate and not need zeroing (but must zero in off-chance of slippage on upper sprockets). All references in RobotContainer to internal enc was switched to external. Other than stated above, external encoder should function the same as internal.