Updated auto-align to consider elevator height.
Moved from scaler to cap-off to limit max output. VelocityY (strafe) max output is decreased as the elevator height is increased. The goal is to minimize steady-state error by moving away from previous scaler, and prevent tippiness when elevator is articulated while auto-aligning. Theory is that limiting overall velocity will reduce the need to limit the acceleration (due to low delta in velocity).