Elevator feedbackSensor fix
Swapped leader and follower so that leader is the one connected to external encoder (and swapped where we get the bottom limit switch and external encoder accordingly), and added external encoder as feedbackDevice in motor config for slot0 & slot1. Reduced all PID and maxMotion constants by the difference between internal rotations to external rotations to preserve the same control we had while using the internal encoder. Renamed some constants for readability. Added descriptive comments throughout. Tanked currentLimit to 30 for initial testing, and added print-outs to confirm corrects slots are being called when we set elevator PID controller reference to a position.